[xsarm_dual_joy] Adds xsarm_dual_joy pkg
This package allows the user to control two X-Series arms at the same
time using a single controller. This package relies on the xsarm_dual
and the xsarm_joy packages for this functionality. The package conists
of a single launch file that configures the two lower-level packages,
allowing for the dual joystick control.